requirements.txt
attrdict==2.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
colored==1.4.4
et-xmlfile==1.1.0
Gooey==1.0.8.1
idna==3.4
numpy==1.24.0
opencv-python==4.6.0.66
openpyxl==3.0.10
pandas==1.5.2
peutils==0.0.44
Pillow==9.3.0
psutil==5.9.4
pygtrie==2.5.0
python-dateutil==2.8.2
pytz==2022.7
requests==2.28.1
scipy==1.9.3
shapely==2.0.0
six==1.16.0
urllib3==1.26.13
wxPython==4.1.0
xmltodict==0.13.0
